Refactor; moved grouped components into group, pulled bio edit state into own component
